feat(LatestPapersView.tsx): add loading skeleton for latest papers view
The imports in the file have been organized and unused imports have been removed to improve code readability. A loading skeleton has been added to the latest papers view to provide a visual indication of loading state. Additionally, a link to read more publications has been added to the bottom of the view to allow users to navigate to the publication page.