feat(PublicationsView.tsx): add PublicationsView component to display a list of publications
The PublicationsView component is added to the project. It receives an array of PublicationItem objects as a prop and renders a container with a header displaying "Publications" and a list of SinglePublication components for each item in the data array. The component also includes styling using Material-UI's makeStyles hook. This component is created to provide a visual representation of a list of publications on the front page of the application.