feat(publication): add new page for publications
A new page for publications has been added to the application. This page imports the `Publications` component from the `Publication` feature and exports it as the default export. Additionally, the `access` constant is exported as `ACCESS.public` from the `routes/types` module. This new page allows users to view publications and sets the access level to public.