refactor(PublicationsView.tsx): update styling in PublicationsView component
The padding in the container class has been changed to "0px 6rem 1rem 6rem" to provide more space around the content. The background color has been commented out for now. The bottomLink class has been removed as it is no longer used. The spacing in the Grid component has been changed to 2 to increase the gap between items. The component prop in the Grid component has been changed to "ol" to use an ordered list instead of an unordered list. The title in the header has been changed to "Latest Publications" for better clarity.