feat(AddPageButton.tsx): add AddPageButton component
The AddPageButton component is added to the project. This component is a button that links to the addpage route and passes the current slug as a state. It is styled using Material-UI and has a primary color variant. This component provides a way for users to add a new page to the application.