feat(AddPageButton.tsx): update button to use Link component from react-router-dom and set the "to" prop to "../addpage" to navigate to the /addpage route
The unused import of useNavigate and the onClick function have been removed as they are no longer needed. The button now uses the Link component from react-router-dom to navigate to the /addpage route when clicked. The "to" prop of the Link component is set to "../addpage" to navigate relative to the current path.