feat(ui-common): add ContentView component for rendering editable content pages
The ContentView component is a React component that renders a view for editable content pages. It takes in a `data` prop which contains the content to be displayed. The component renders a Container from Material-UI and an Editor component from the "editor" package. The Editor component is passed the content to be displayed and is set to be non-editable. This component is useful for rendering content pages that need to be displayed in a read-only format.