fix(Dockerfile.stock-center): add --ignore-engines flag to yarn install command to avoid engine compatibility issues
The --ignore-engines flag is added to the yarn install command in the
Dockerfile to prevent the build process from failing due to engine
compatibility issues. This change ensures that the Docker image can be
built even if the Node.js version specified in the package.json file
does not match the version installed in the Docker image.