feat(ui-common): move EditView component for editing content
The EditView component is moved to the ui-common package. This component is responsible for rendering an editable view of content. It receives data as a prop, which includes the content to be edited. It uses the Editor component from the "editor" package to provide editing functionality. The onSave function is called when the user saves the changes, and the onCancel function is called when the user cancels the editing. The navigate function from react-router-dom is used to navigate back to the editable view after saving or canceling.