feat(ci.yml): add new job for building and pushing Docker image for frontpage
The Node.js version used in the GitHub Actions workflow is upgraded to
20.x to take advantage of the latest features and improvements in the
runtime. A new job is added to the workflow to build and push a Docker
image for the frontpage. This allows the frontpage to be deployed as a
Docker container, which can improve scalability and ease of deployment.