feat(auth/package.json): add @dictybase/ui-common as a dependency
The @dictybase/ui-common package was added as a dependency in the auth
package to provide common UI components across different packages. This
led to the refactoring of the import statement in Private.tsx, replacing
@dictybase/ui-dsc with @dictybase/ui-common for LoadingDisplay. This
change ensures that all UI components are imported from the same
package, improving consistency and maintainability. The nanoid
dependency was removed from the news-component package as it was no
longer being used, reducing the package's size and complexity.