fix(ci.yml): use --ignore-engines flag in yarn install command to avoid engine check
The --ignore-engines flag is added to the yarn install command in the CI
workflow to prevent the build from failing due to engine version
mismatches. This allows the build to proceed even if the Node.js version
specified in the package.json's "engines" field does not match the one
used in the CI environment.