fix(dicty-frontpage): change base path in vite.config.ts from "/stockcenter" to "/"
The DEPLOY_ENV variable was removed from .env.development as it was no
longer needed in the development environment. The base path in
vite.config.ts was changed to "/" to ensure that the application is
served from the root path, which simplifies the routing structure and
makes the application more accessible.