feat(EditView.tsx): add functionality to update content using useUpdateContentMutation hook
The EditView component now uses the useUpdateContentMutation hook from the dicty-graphql-schema package to update the content. The updateContent function is called when the onSave function is triggered, passing the updated content and other necessary variables. This allows the user to save changes made in the editor and update the content in the database.