refactor(AddPage.tsx): refactor AddPage component
- Import `EditableContentLoadingDisplay` from `@dictybase/ui-common` and remove unused import `Loader`
- Remove unused variables `gqlLoading` and `logToLoading`
- Replace `<Loader />` with `<EditableContentLoadingDisplay />` when `loading` is true
- Simplify pattern matching for `data` object in `useContentBySlugQuery` hook
- Remove unnecessary condition for `hasNotFoundError` check in `when` block

The changes were made to improve code readability and remove unused code. The `EditableContentLoadingDisplay` component is now used instead of the `Loader` component for displaying loading state. The destructuring of hooks and variables has been simplified for better code organization. The condition for `hasNotFoundError` check in the `when` block has been simplified.
Closes #648