feat(EditView.tsx): add createToolbarWrapper function import and implement toolbar component
The eslint-disable comment was added to ignore linting errors related to camelcase rule for consistency with existing code style. The createToolbarWrapper function was imported and used to create a toolbar component in the EditView component, displaying information such as the last update time and the name of the user who made the update.