feat(EmptyPublications.tsx): add EmptyPublications component to display a message when there are no publications for a given range
The EmptyPublications component is added to display a message when there are no publications available for a given range. It renders a container with a fixed height and a grey background color. Inside the container, it displays a centered message with the range for which there are no publications. This component improves the user experience by providing feedback when there are no publications to display.