feat(PublicationsList.tsx): add PublicationsList component to display a list of publications based on the current tab
The PublicationsList component is added to the project. It takes in two props: sortedPublications, which is an object containing sorted publications grouped by tab name, and currentTab, which is the name of the current tab being displayed. The component renders a list of publications based on the current tab, with each publication being displayed using the SinglePublication component. If there are no publications for the current tab, the EmptyPublications component is rendered instead. The component also applies styling using the makeStyles hook.