feat(PublicationLinks.tsx): add PublicationLinks component to display links to publication archives
The PublicationLinks component is added to display links to publication archives based on the digital identifiers associated with a publication. The getIdentifierUrl function is used to generate the appropriate URL for each identifier. The component takes an array of identifiers as a prop and renders a grid of links to the publication archives.