feat(ui-common): add declaration file for image assets and add sad-dicty.png image asset
The commit adds a declaration file assets.d.ts to the ui-common package, which declares module "*.png" and "*.jpg" to allow importing image assets with those file extensions. Additionally, the commit adds the sad-dicty.png image asset to the src/assets directory in the ui-common package. This image asset can now be used in the application.