feat(show.tsx): add ShowHomepage component for the main homepage of DSC
The ShowHomepage component is added to serve as the main homepage for the DSC (Stock Center) application. It imports various UI components from the "@dictybase/ui-dsc" and "@dictybase/ui-common" packages to build the homepage layout. The component also uses the "useContentBySlugQuery" hook from the "dicty-graphql-schema" package to fetch content data based on a specific slug. The component handles different scenarios such as loading, error, and successful data retrieval to display the appropriate UI components. The component is exported as the default export and also exports an "access" constant for public access.