chore(jest.config.js): update moduleNameMapper for "react-markdown" to point to the correct file extension
The moduleNameMapper in the jest.config.js file has been updated to point to the correct file extension for the "react-markdown" module. It was previously pointing to a .js file, but it has been changed to point to a .tsx file. This change ensures that the correct file is used for mocking the "react-markdown" module during testing.