feat(proteininformation.tsx): add ProteinInformationPageWrapper component to render the Protein Information page for a given gene id
The ProteinInformationPageWrapper component is added to render the Protein Information page for a specific gene id. It uses the useRouter hook from Next.js to get the gene id from the query parameters. It then uses the useGeneQuery hook from the dicty-graphql-schema package to fetch the gene data. The component conditionally renders a loading spinner, an error page if there is an error, or the ProteinInfoContainer component if the data is available.