chore: handle react/no-array-index-key errors
In many React components, array indices were improperly used as keys. In these cases, a more suitable, unique value of the array item was used as a key instead. In the case of loaders or other components that had no unique identifies, the rule was disabled for that line or file.