refactor(types.ts): add support for Plasmid type and create CatalogItem type to represent both StrainItem and PlasmidItem
The types.ts file has been modified to import the Plasmid type from "dicty-graphql-schema" and include it in the existing types. The StrainItem type has been updated to include the "__typename" field from the Strain type. Additionally, a new PlasmidItem type has been created to represent the fields of the Plasmid type. To handle both StrainItem and PlasmidItem, a new CatalogItem type has been introduced. This change allows for more flexibility when working with different types of items in the catalog.